requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
boto3==1.29.7
python-dotenv==1.0.0
Pillow==10.0.1